refactor(drop-vector-index): resolve segment transformers from a global registry, drop the bucket option
