+++ adapters/repos/db/lsmkv/compactor_replace_transform_test.go
@@ -38,9 +38,9 @@ func newReplaceBucketForCompaction(t *testing.T, transformer valueTransformer) *
 	// Inject the transformer via the edit-ops path: a registered op makes
 	// BuildCurrentTransformer dispatch to this type's factory per pass.
 	if transformer != nil {
-		editOps := newSegmentEditOps(bucket.disk.dir, map[OpType]OpTransformerFactory{
-			OpTypeRemoveTargetVectors: func(ops []ActiveOp) func([]byte) ([]byte, error) { return transformer },
-		})
+		editOps := newSegmentEditOpsWithLookup(bucket.disk.dir, "TestClass", false, staticResolver(map[OpType]OpTransformerFactory{
+			OpTypeRemoveTargetVectors: func(className string, skip bool, ops []ActiveOp) func([]byte) ([]byte, error) { return transformer },
+		}))
 		t.Cleanup(func() { require.NoError(t, editOps.Close()) })
 		bucket.disk.editOps = editOps
 		require.NoError(t, editOps.RegisterOp("test-op",
