+++ adapters/repos/db/lsmkv/segment_group.go
@@ -141,7 +141,8 @@ type sgConfig struct {
 	writeMetadata                bool
 	sequentialAccess             bool
 	shouldSkipKey                func(key []byte, ctx context.Context) (bool, error)
-	editOpTransformers           map[OpType]OpTransformerFactory
+	className                    string
+	skipClassNameOnDisk          bool
 }
 
 func newSegmentGroup(ctx context.Context, logger logrus.FieldLogger, metrics *Metrics, cfg sgConfig,
@@ -544,8 +545,12 @@ func newSegmentGroup(ctx context.Context, logger logrus.FieldLogger, metrics *Me
 	// published before any pass can read it (happens-before). The bolt file itself
 	// is opened lazily on the first registered op (see newSegmentEditOps), so an
 	// objects bucket that never sees a drop carries no sidecar. Closed in shutdown.
-	if len(cfg.editOpTransformers) > 0 {
-		sg.editOps = newSegmentEditOps(cfg.dir, cfg.editOpTransformers)
+	//
+	// A non-empty className means the objects bucket (the only WithClassName caller);
+	// edit ops only apply to its replace-strategy store. Transformers are resolved
+	// per op type from the global registry; the persisted ops drive what runs.
+	if cfg.className != "" && cfg.strategy == StrategyReplace {
+		sg.editOps = newSegmentEditOps(cfg.dir, cfg.className, cfg.skipClassNameOnDisk)
 		sg.editOps.logger = sg.logger
 	}
