+++ adapters/repos/db/shard_write_drop_vector.go
@@ -14,40 +14,10 @@ package db
 import (
 	"fmt"
 
-	"github.com/weaviate/weaviate/adapters/repos/db/lsmkv"
 	"github.com/weaviate/weaviate/entities/models"
 	"github.com/weaviate/weaviate/entities/modelsext"
 	"github.com/weaviate/weaviate/entities/storobj"
 )
-
-// dropVectorTransformer is the OpTransformerFactory for OpTypeRemoveTargetVectors:
-// given the live remove-target-vector ops it strips the named vectors from each
-// stored object, re-marshaling only when something changed (idempotent no-op
-// otherwise). Every op passed in is already of this type — the edit-ops registry
-// dispatches by type — so no type check is needed here. The disk round-trip must
-// match the bucket's own encode/decode (FromBinaryDisk / MarshalBinaryDisk).
-func dropVectorTransformer(className string, skipClassNameOnDisk bool) lsmkv.OpTransformerFactory {
-	return func(ops []lsmkv.ActiveOp) func([]byte) ([]byte, error) {
-		return func(value []byte) ([]byte, error) {
-			obj, err := storobj.FromBinaryDisk(value, className)
-			if err != nil {
-				return nil, fmt.Errorf("decode object for vector drop: %w", err)
-			}
-			changed := false
-			for _, op := range ops {
-				for _, targetVector := range op.Descriptor.Targets {
-					if obj.RemoveTargetVector(targetVector) {
-						changed = true
-					}
-				}
-			}
-			if !changed {
-				return value, nil
-			}
-			return obj.MarshalBinaryDisk(skipClassNameOnDisk)
-		}
-	}
-}
 
 // isDroppedVectorIndex reports whether the named vector's index has been
 // dropped (VectorConfig entry marked VectorIndexType=="none"). The legacy
